+++ CHANGELOG.md
@@ -24,6 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - [api] Do not attempt to load config from core in ephemeral mode
+- [api] Fix loading of info from core in ephemeral mode
 - [core] /jwks endpoint in ephemeral mode
 - [core] Typo in returned body
 - [ui] MRT login was sending incorrect data